requirements.txt
protobuf==5.28.3
python-dotenv==1.0.1
streamlit==1.39.0
